:recycle: Refactor connection factory into a class
